401. Transforming "Then Catch" to "Async Await"

+++ controllers/feed.js
@@ -1,36 +1,29 @@
 const {validationResult} = require('express-validator')
 const Post = require('../models/post')
+const User = require('../models/user')
 const fs = require('fs')
 const path = require("path");
 
-exports.getPosts = (req, res, next) => {
+exports.getPosts = async (req, res, next) => {
     const currentPage = req.query.page || 1;
     const perPage = 2;
-    let totalItems;
-    Post.find()
-        .countDocuments()
-        .then(count => {
-            totalItems = count
-            return Post.find()
-                .skip((currentPage-1) * perPage)
-                .limit(perPage)
-        })
-        .then(posts => {
-            res.status(200)
-                .json({
-                    message:'Fetched Posts Successfully',
-                    posts,
-                    totalItems
-                })
-        })
-        .catch(err =>{
-            if (!err.statusCode) {
-                err.statusCode = 500;
-            }
-            next(err)
-        })
 
+    try {
+        const totalItems = await Post.find().countDocuments();
+        const posts = await Post.find()
+            .skip((currentPage - 1) * perPage)
+            .limit(perPage)
 
+        res.status(200)
+            .json({
+                message: 'Fetched Posts Successfully', posts, totalItems
+            })
+    } catch (err) {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err)
+    }
 }
 
 exports.postPosts = (req, res, next) => {
@@ -48,15 +41,26 @@ exports.postPosts = (req, res, next) => {
     }
     const imageUrl = req.file.path;
     const {title, content} = req.body;
-
+    let creator;
     const post = new Post({
-        title, content, imageUrl: imageUrl, creator: {name: 'Alvin'}
+        title, content, imageUrl, creator: req.userId
     });
     post.save()
         .then(result => {
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            creator = user
+            user.posts.push(post)
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
-                message: 'Post created successfully!', post: result
+                message: 'Post created successfully!', post: post, creator: {
+                    _id: creator._id, name: creator.name
+                }
             })
+
         })
         .catch(err => {
             if (!err.statusCode) {
@@ -76,8 +80,7 @@ exports.getPost = ((req, res, next) => {
                 throw error;
             }
             res.status(200).json({
-                message: 'Post fetched',
-                post
+                message: 'Post fetched', post
             })
         })
         .catch(err => {
@@ -108,6 +111,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized.')
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl)
             }
@@ -118,8 +126,7 @@ exports.updatePost = (req, res, next) => {
         })
         .then(result => {
             res.status(200).json({
-                message: 'Post Updated!',
-                post: result
+                message: 'Post Updated!', post: result
             })
         })
         .catch(err => {
@@ -139,11 +146,23 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not Authorized.')
+                error.statusCode = 403;
+                throw error;
+            }
             clearImage(post.imageUrl)
             return Post.findByIdAndRemove(postId);
         })
         .then(result => {
-            console.log(result)
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            user.posts.pull(postId)
+            return user.save();
+        })
+        .then(result => {
+            // console.log(result)
             res.status(200).json({
                 message: 'Deleted Post!'
             })
